Add fixed dimensions to image render
I've tried to explain the origin of these numbers. they mostly come from
trial an error with a bit of educated guessing.

the images themselves are square 1:1, but that includes transparent padding.

width, while it's not used to constrain the image, to account for
the aspect ratio of monospace characters, should be set 1.5-2x the height
as far as i can determine by comparing a few monospace fonts:
e.g Monaco, Menlo